Continued work.   Please review as these are a LOT of manual changes.   Compiles and you can log in.

+++ Source/ACE/Entity/Landblock.cs
@@ -640,7 +640,7 @@ namespace ACE.Entity
                             if ((aPlayer != null) && (inventoryItem != null))
                             {
                                 if (aPlayer.PhysicsData.Position.SquaredDistanceTo(inventoryItem.PhysicsData.Position)
-                                    > Math.Pow(inventoryItem.GameData.UseRadius, 2))
+                                    > Math.Pow((double)inventoryItem.GameData.UseRadius, 2))
                                 {
                                     // This is where I need to hook in the move to object code.
                                     // TODO: Og II work on this soon.
@@ -658,7 +658,7 @@ namespace ACE.Entity
                                 motion = new UniversalMotion(MotionStance.Standing);
                                 aPlayer.Session.Network.EnqueueSend(new GameMessagePrivateUpdatePropertyInt(aPlayer.Session,
                                        PropertyInt.EncumbranceVal,
-                                       aPlayer.GameData.Burden),
+                                       (uint)aPlayer.GameData.Burden),
                                        new GameMessagePutObjectInContainer(aPlayer.Session, aPlayer, inventoryId),
                                        new GameMessageUpdateMotion(aPlayer, aPlayer.Session, motion),
                                        new GameMessageUpdateInstanceId(inventoryId, playerId),
